+++ src/config/blog.js
@@ -2,5 +2,6 @@ export const constants = {
 	updateInterval: 30, // seconds between each fetch
 	totalPosts: 5, // Number of posts to cycle through
 	host: 'http://elsys.blog/',
-	basePath: 'wp-json/wp/v2/'
+	basePath: 'wp-json/wp/v2/',
+	summaryFallbackToExcerpt: true // set this to false if not excerpt should be used when no summary
 }